+++ index.ts
@@ -2,6 +2,7 @@ export * from "./eosiolib/asset";
 export * from "./eosiolib/check";
 export * from "./eosiolib/symbol";
 export * from "./eosiolib/symbol_code";
+export * from "./eosiolib/extended_symbol";
 export * from "./eosiolib/time";
 export * from "./eosiolib/name";
 export * from "./eosiolib/utils";